mem0_memory_test_app/backend/requirements.txt
# Mem0 + vector stores
mem0ai>=2.0.0
langchain-voyageai>=0.1.0
pymongo>=4.0.0
pgvector>=0.2.0
psycopg2-binary>=2.9.0

# API & server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# OpenAI (Azure)
openai>=1.0.0

# Env
python-dotenv>=1.0.0
